introduce WEBGL_PACK_IMAGE_OPERATIONS
