+++ integration_tests/benchmarks/reduction_ops_benchmark.ts
@@ -73,53 +73,3 @@ export class ReductionOpsGPUBenchmark implements BenchmarkTest {
     return time;
   }
 }
-
-export class AvgPoolGPUBenchmark implements BenchmarkTest {
-  async run(size: number) {
-    tf.setBackend('webgl');
-
-    const x: tf.Tensor4D = tf.randomUniform([1, size, size, 320], -1, 1);
-
-    const benchmark = () => tf.avgPool(x, [size, size], [size, size], 'valid');
-
-    const time = await util.warmupAndBenchmarkGPU(benchmark);
-
-    x.dispose();
-
-    return time;
-  }
-}
-
-export class ConcatGPUBenchmark implements BenchmarkTest {
-  async run(size: number) {
-    tf.setBackend('webgl');
-
-    const x: tf.Tensor4D = tf.randomUniform([1, size, size, 256], -1, 1);
-    const y: tf.Tensor4D = tf.randomUniform([1, size, size, 256], -1, 1);
-
-    const benchmark = () => tf.concat([x, y], 3);
-
-    const time = await util.warmupAndBenchmarkGPU(benchmark);
-
-    x.dispose();
-    y.dispose();
-
-    return time;
-  }
-}
-
-export class ResizeBilinearBenchmark implements BenchmarkTest {
-  async run(size: number) {
-    tf.setBackend('webgl');
-
-    const x: tf.Tensor4D = tf.randomUniform([1, size, size, 21], -1, 1);
-
-    const toSize = (size - 1) * 8 + 1;
-    const benchmark = () => tf.image.resizeBilinear(x, [toSize, toSize], true);
-
-    const time = await util.warmupAndBenchmarkGPU(benchmark);
-
-    x.dispose();
-    return time;
-  }
-}
