+++ src/environment.ts
@@ -318,7 +318,7 @@ export class Environment {
       return this.get('WEBGL_PACK');
     } else if (feature === 'WEBGL_PACK_BINARY_OPERATIONS') {
       return this.get('WEBGL_PACK');
-    } else if (feature === 'WEBGL_PACK_ARRAY_OPERATIONS') {
+    } else if (feature === 'WEBGL_PACK_IMAGE_OPERATIONS') {
       return this.get('WEBGL_PACK');
     } else if (feature === 'WEBGL_LAZILY_UNPACK') {
       return this.get('WEBGL_PACK');
