+++ src/environment_util.ts
@@ -40,6 +40,8 @@ export interface Features {
   'WEBGL_PACK_BINARY_OPERATIONS'?: boolean;
   // Whether we will pack SpaceToBatchND, BatchToSpaceND, slice, pad, transpose.
   'WEBGL_PACK_ARRAY_OPERATIONS'?: boolean;
+  // Whether we will pack image operations.
+  'WEBGL_PACK_IMAGE_OPERATIONS'?: boolean;
   // Whether we will use the im2col algorithm to speed up convolutions.
   'WEBGL_CONV_IM2COL'?: boolean;
   // The maximum texture dimension.
@@ -113,6 +115,7 @@ export const URL_PROPERTIES: URLProperty[] = [
   {name: 'WEBGL_PACK_DEPTHWISECONV', type: Type.BOOLEAN},
   {name: 'WEBGL_PACK_BINARY_OPERATIONS', type: Type.BOOLEAN},
   {name: 'WEBGL_PACK_ARRAY_OPERATIONS', type: Type.BOOLEAN},
+  {name: 'WEBGL_PACK_IMAGE_OPERATIONS', type: Type.BOOLEAN},
   {name: 'WEBGL_CONV_IM2COL', type: Type.BOOLEAN},
   {name: 'WEBGL_MAX_TEXTURE_SIZE', type: Type.NUMBER},
   {name: 'WEBGL_NUM_MB_BEFORE_PAGING', type: Type.NUMBER},
